The beginning
The beginning of adding new commands to the checkIsCommand.
first command stop command.
(not in anyway finished jsut started)
instead of in check checking for a wall it checks for stop.
that way anything can stop not just the wall.

still to do for this.
refactor checkIsCommand to use the new this sttributes.